Crate casm

Reexports

pub use errors::*;

Modules

codec

config

errors

musicfile

transcoder

Statics

ALL_CODECS

Functions

process_files
Runs the processing and copying on each file. This is done in a multithreaded manner and shows a progressbar.

run
Runs the program

scan_files
Creates a unique set of music files from the configuration data. It expands globs and walks directories