Struct casm::codec::Codec

pub struct Codec<'a> {
    pub name: &'a str,
    pub lossless: bool,
    pub extension: &'a str,
}

A codec type for storing basic metadata used for conversion later.

Fields
name: &'a str

The name of the codec, as specified in enum ffmpeg::codec::id::Id

lossless: bool

Whether the codec is lossless or lossy

extension: &'a str

The extension of the codec (i.e. .mp3)

Methods

impl<'a> Codec<'a>

fn is_acceptable(&self, acceptable_formats: &[String]) -> bool

Checks to see if the codec is one of the acceptable codecs as specified in the user configuration

Arguments

'acceptable_formats' - The list of acceptable formats provided in the config