Struct casm::config::Config

pub struct Config {
    pub source_folder: String,
    pub dest_folder: String,
    pub exclude: Option<Vec<String>>,
    pub files: Vec<String>,
    pub convert_profile: ConvertProfile,
}

The user-specified config

Fields
source_folder: String

The source folder that all files are relative to (i.e. ~/Music/)

dest_folder: String

The dest folder to which all relative filenames will be appended (i.e. /mnt/Internal_Storage/Music)

exclude: Option<Vec<String>>

The regex that matches files to exclude, usually for instrumental tracks

files: Vec<String>

A list of files, folders, and glob patterns to convert (this will be unique'd later)

convert_profile: ConvertProfile

Conversion-specific settings

Methods

impl Config

fn new(config_cli: Option<&str>) -> Result<Config>

Creates a config struct from the configuration file

Arguments

config_cli - An Option that may hold the location of the configuration file. Otherwise, the xdg location will be used.