Struct casm::config::ConvertProfile

pub struct ConvertProfile {
    pub target_format: String,
    pub acceptable_formats: Vec<String>,
    pub bit_rate: usize,
}

The conversion-specific parts of the user-specified config

Fields
target_format: String

A ffmpeg::codec::id::Id that is supported (see gen_codec_types). Files that are not already one of acceptable_formats will be converted to this format.

acceptable_formats: Vec<String>

A list of music types that are acceptable on the target. If the source file is one of these, it will not be converted to target_format. The list may include types: * ffmpeg::codec::id::Ids that are supported (see gen_codec_types) * lossless for all lossless files (not recommended, unless also combined with lossy) * lossy for all lossy files (recommended) I usually go with lossy here, so that I'm not converting between lossy formats (more loss!) but don't have giant lossy files on my phone.

bit_rate: usize

A target bit rate in KB/s (i.e. 320 or 128)