Function casm::transcoder::select_best_rate

fn select_best_rate(source: i32, supported: Option<RateIter>) -> i32

Get the closest sample rate without going under, unless there is nothing above There are some issues with this. Sample rate conversion can have some loss, and we're letting the user target a bitrate. That said, we're mostly doing this to convert from 44.1khz to 48khz for opus.

This isn't as fast as it could be, but that's ok. The vectors will be small (<=10).

Arguments

source - The sample rate of the source file

supported - The supported sample rates of the target codec